# perf(staking): optimize endblock by reducing bech32 conversions