This is a rebased version of #23869 with some minor tweak and benchmark. All credits and bit thanks go to @ValarDragon for the original patch and idea!


### Performance gains in the test setup to original code

Map key patch
`((654391 - 435196) / 654391) * 100 ~= 33 %`

Map value patch
`((654391 - 421474) / 654391) * 100 ~= 35 %`


### Average

Original 
`(648741 + 649484 + 667771 + 651568) / 4 = 654391 ns/op`
Map key patch
`(433791 + 432525 + 432654 + 441817) / 4 =  435196 ns/op`
Map value patch
`(421966 + 422146 + 419467 + 422317) / 4 =  421474 ns/op`

### Benchmarks on my box

* Original code
```
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    1789	    648741 ns/op
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    1808	    649484 ns/op
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    1755	    667771 ns/op
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    1792	    651568 ns/op
```

* With ValarDragon's patch to use non bech32 map key
```
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    2646	    433791 ns/op
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    2608	    432525 ns/op
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    2625	    432654 ns/op
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    2661	    441817 ns/op
```

* With my patch using int64 map value

```
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    2768	    421966 ns/op
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    2654	    422146 ns/op
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    2752	    419467 ns/op
BenchmarkApplyAndReturnValidatorSetUpdates-12    	    2703	    422317 ns/op




I did a quick microbenchmark in the beginning for marshal/unmarshal of `gogotypes.Int64Value` type. The unmarshal is ~41% faster so I refactored the index to use stdlib int64 instead. In the `BenchmarkApplyAndReturnValidatorSetUpdates` this brought ~2% performance gain.

```
BenchmarkMustMarshal
BenchmarkMustMarshal-12      	60243637	        19.92 ns/op
BenchmarkMustUnmarshal
BenchmarkMustUnmarshal-12    	100000000	        11.69 ns/op
```

<!-- Add a description of the changes that this PR introduces and the files that
are the most critical to review. -->

---

## Author Checklist

*All items are required. Please add a note to the item if the item is not applicable and
please add links to any relevant follow up issues.*

I have...

* [ ] included the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title, you can find examples of the prefixes below:
    <!-- * `feat`: A new feature
    * `fix`: A bug fix
    * `docs`: Documentation only changes
    * `style`: Changes that do not affect the meaning of the code (white-space, formatting, missing semi-colons, etc)
    * `refactor`: A code change that neither fixes a bug nor adds a feature
    * `perf`: A code change that improves performance
    * `test`: Adding missing tests or correcting existing tests
    * `build`: Changes that affect the build system or external dependencies (example scopes: gulp, broccoli, npm)
    * `ci`: Changes to our CI configuration files and scripts (example scopes: Travis, Circle, BrowserStack, SauceLabs)
    * `chore`: Other changes that don't modify src or test files
    * `revert`: Reverts a previous commit -->
* [ ] confirmed `!` in the type prefix if API or client breaking change
* [ ] targeted the correct branch (see [PR Targeting](https://github.com/cosmos/cosmos-sdk/blob/main/CONTRIBUTING.md#pr-targeting))
* [ ] provided a link to the relevant issue or specification
* [ ] reviewed "Files changed" and left comments if necessary
* [ ] included the necessary unit and integration [tests](https://github.com/cosmos/cosmos-sdk/blob/main/CONTRIBUTING.md#testing)
* [ ] added a changelog entry to `CHANGELOG.md`
* [ ] updated the relevant documentation or specification, including comments for [documenting Go code](https://blog.golang.org/godoc)
* [ ] confirmed all CI checks have passed

## Reviewers Checklist

*All items are required. Please add a note if the item is not applicable and please add
your handle next to the items reviewed if you only reviewed selected items.*

Please see [Pull Request Reviewer section in the contributing guide](../CONTRIBUTING.md#reviewer) for more information on how to review a pull request.

I have...

* [ ] confirmed the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
* [ ] confirmed all author checklist items have been addressed
* [ ] reviewed state machine logic, API design and naming, documentation is accurate, tests and test coverage


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **Bug Fixes**
  - Prevented unexpected application interruptions in staking operations by ensuring clear, descriptive error messages during failures.

- **Refactor**
  - Streamlined the handling of validator state transitions and power updates for enhanced reliability and robustness.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->